Implement framebuffer using libremarkable
The refresh modes are not fully tested yet as inputs are
not implemented right now. Homescreen looks nice though.

Since the last port of plato the the libremarkable lib
seemed to be finished. There are also some new update
optimizations in plato that require a yet to be made
pull request to libremarkable to get working with
maybe even better latency.