Tune refresh modes
The refreshmodes are the result of a lot of hours trying
a lot of stuff and setteling for that. There doesn't seem
to be a magic mode or similar. Currently not really suited
for manga reading but works good for epubs.

TODO: The UI seems to be unstable (cause unknown). Sometimes
after a lot of switching, the UI will stop responding for
certain actions or some parts of the UI won't be clickable
anymore.